% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_trans}
\alias{date_trans}
\title{Apply transformation from Wiki text to date}
\usage{
date_trans(x)
}
\arguments{
\item{x}{vector with parsed raw date text}
}
\description{
Apply transformation from Wiki text to date
}
