% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamagabinetes-package.R
\docType{package}
\name{jamagabinetes-package}
\alias{jamagabinetes}
\alias{jamagabinetes-package}
\title{jamagabinetes: Visualización de Gabinetes}
\description{
El paquete descarga las líneas de tiempo de todos los puestos de
  gabinete de Wikipedia y las visualiza a través del tiempo.
}
\author{
\strong{Maintainer}: Juan Hernandez \email{jamamel@gmail.com}

}
\keyword{internal}
