% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_scrape.R
\name{scrape_cabinet}
\alias{scrape_cabinet}
\title{Scrape Wiki page for cabinet data and produce a tidy data.table for analysis}
\usage{
scrape_cabinet(url)
}
\arguments{
\item{url}{url to wiki page from which to extract cabinet information}
}
\description{
Scrape Wiki page for cabinet data and produce a tidy data.table for analysis
}
\examples{
url <- xml2::read_html('https://es.wikipedia.org/wiki/Anexo:Gabinete_de_Benito_Ju\%C3\%A1rez')
parsed_dt <- scrape_cabinet(url)

}
